backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.9
supabase==2.5.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-groq==0.1.3
tavily-python==0.3.3
httpx==0.27.0
langchain-huggingface==0.0.3
unstructured[docx,pdf]==0.14.5
faiss-cpu==1.8.0

